+++ examples/scripts/comparison_examples/nelder_mead.py
@@ -22,10 +22,8 @@ parameters = pybop.Parameters(
 sigma = 0.003
 experiment = pybop.Experiment(
     [
-        (
-            "Discharge at 0.5C for 3 minutes (3 second period)",
-            "Charge at 0.5C for 3 minutes (3 second period)",
-        ),
+        "Discharge at 0.5C for 3 minutes (3 second period)",
+        "Charge at 0.5C for 3 minutes (3 second period)",
     ]
     * 2
 )